requirements.txt
PyQt6
pyserial
numpy